+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Version {next}
   - Better tag normalization code organisation
+  - Remove `jonnybarnes/unicode-tools` dependency and clean up relevant code
 
 ## Version 0.0.2 (2016-05-25)
   - Fix issue#1: tagged notes page needs the tag from the URL normalizing.